// ticket-manager frent end/src/app/invoices/models/invoice.ts
import { Client } from '../../clients/models/client';

export class Invoice {
	_id: string;
	item: string;
	qty: number;
	date: Date;
	due: Date;
	tax: number;
	rate: number;
	price: number;
	client: Client;
}
export class InvoicePaginationRsp {
	docs: Invoice[];
	total: number;
	pages: number;
	page: number;
	limit: number;
}